Corrected near pipecell cells allocation to edges and fixed a bug in protocol secretion values
It was hard for the algorithm to allocate near pipes cells to some edges
due to high distance. I put a range of closeness to pipe cells to
overweigh distance issue so that those cells get assigned to correct edges
and get their correct secretion rates.
Also, the secretion rates did not get correctly updated in protocol.xml
file due to a small but which was fixed here.